chore: rename migration head to codegen head
