+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/codegen-head/amplify_stack_parser.ts
