+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/codegen-head/analytics.ts
