+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/codegen-head/directory_exists.ts
