+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/codegen-head/extract_sub_commands.ts
